requirements.txt
Pillow==9.1.1
albumentations==1.2.1
tqdm==4.64.0
prefetch_generator==1.0.1
